requirements.txt
# AI Research Orchestrator - Dependencies

# OpenAI Agents SDK
openai-agents>=0.1.0
openai>=1.50.0

# Google Gemini
google-generativeai>=0.8.0

# Web Search & HTTP
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
trafilatura>=1.8.0

# Code Execution (sandboxed)
docker>=7.0.0
RestrictedPython>=7.0

# Async Support
asyncio>=3.4.3
anyio>=4.0.0

# Data Processing
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
tenacity>=8.2.0
structlog>=24.1.0

# Type hints
typing-extensions>=4.9.0
